Make a package for test base classes.
