+++ tests/test_feedexport_batch.py
@@ -21,8 +21,9 @@ from scrapy.settings import Settings
 from scrapy.utils.python import to_unicode
 from scrapy.utils.test import get_crawler
 from tests.spiders import ItemSpider
+from tests.utils.bases.feedexport import TestFeedExportBase
 from tests.utils.decorators import coroutine_test, inline_callbacks_test
-from tests.utils.feedexport import MyItem, TestFeedExportBase
+from tests.utils.feedexport import MyItem
 
 if TYPE_CHECKING:
     from os import PathLike
